src/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.2
django-admin-csvexport==1.9
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.1
idna==2.10
importlib-metadata==3.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uwsgi==2.0.19.1
whitenoise==5.2.0
zipp==3.4.1
